+++ src/Couchbase.EntityFrameworkCore/Metadata/Conventions/CouchbaseConventionSetBuilder.cs
@@ -19,6 +19,7 @@ public class CouchbaseConventionSetBuilder : RelationalConventionSetBuilder
         conventionSet.Add(new JsonPropertyNameConvention(Dependencies));
         conventionSet.Add(new JsonPropertyConvention(Dependencies));
         conventionSet.Add(new DateTimeFormatConvention(Dependencies));
+        conventionSet.Add(new CouchbaseMetaConvention(Dependencies));
         return conventionSet;
     }
 }